backend/pyproject.toml
[project]
name = "inference-api"
version = "0.1.0"
description = "OpenAI-compatible API for local model inference"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "sse-starlette",
    "torch",
    "transformers>=5.4",
    "accelerate",
    "bitsandbytes",
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "pillow>=12.2.0",
    "torchvision",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
